switch int to int64

+++ pkg/sql/colexec/preinsert/types.go
@@ -38,7 +38,7 @@ type Argument struct {
 	Attrs      []string
 	IsUpdate   bool
 
-	EstimatedRowCount int
+	EstimatedRowCount int64
 
 	buf *batch.Batch
